refactored defaultEvent into sub modules
