+++ src/scripts/index.js
@@ -1,4 +1,4 @@
-const defaultEvent = require('./defaultEvent.js'),
+const defaultEvent = require('./defaultEvent/index.js'),
     meetingEnd = require('./meetingEnd.js')
 
 module.exports = { default: defaultEvent, meetingEnd }
